Now supporting down to Android 4.1 (used to be 4.2)

+++ include/small3d/Model.hpp
@@ -152,7 +152,7 @@ namespace small3d {
      *        table. Each "row" in the index data table forms a triangle.
      *
      */
-    std::vector<uint32_t> indexData;
+    std::vector<uint16_t> indexData;
 
     /**
      * @brief Size of the index data, in bytes
